+++ Tests/DevProjex.Tests.Integration/EmptyFoldersEffectiveDeltaMatrixIntegrationTests.cs
@@ -166,7 +166,7 @@ public sealed class EmptyFoldersEffectiveDeltaMatrixIntegrationTests
 		return count;
 	}
 
-	private static int CountAsciiTreeLines(string rootPath, TreeNodeDescriptor root)
+	private static long CountAsciiTreeLines(string rootPath, TreeNodeDescriptor root)
 	{
 		var treeText = new TreeExportService().BuildFullTree(rootPath, root, TreeTextFormat.Ascii);
 		return ExportOutputMetricsCalculator.FromText(treeText).Lines;
